requirements.txt
# Core Framework
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.4

# Database
SQLAlchemy>=2.0
alembic>=1.12

# Data & ML
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
xgboost>=2.0
joblib>=1.3

# Trading & Market Data
ccxt>=4.2
pybit>=5.6

# News & Data Fetching
feedparser>=6.0
requests>=2.31

# Scheduling & Background Tasks
apscheduler>=3.10

# UI & Reporting
fastapi-offline>=0.3
streamlit>=1.35
jinja2>=3.1
xlsxwriter>=3.1

# Utilities
python-dateutil>=2.8
python-dotenv>=1.0

# Development Tools
ruff>=0.1.9
black>=23.12.0
mypy>=1.8.0
pytest>=7.4
pytest-cov>=4.1